requirements.txt
# LegalInsight - Self-RAG + EigenScore Dependencies

# Core ML & NLP
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
datasets>=2.14.0

# Self-RAG Model Support
llama-cpp-python>=0.2.0
ctransformers>=0.2.27

# Retrieval & Embeddings
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-community>=0.0.10

# Web API
flask>=2.3.0
flask-cors>=4.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Configuration
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

# Optional: For GPU Support
# torch>=2.0.0+cu118 -f https://download.pytorch.org/whl/cu118
# faiss-gpu>=1.7.4

# Development
pytest>=7.4.0
jupyter>=1.0.0
notebook>=7.0.0
